+++ module/GeebyDeeby/src/GeebyDeeby/Controller/UserController.php
@@ -29,7 +29,7 @@
 
 namespace GeebyDeeby\Controller;
 
-use Laminas\Crypt\Password\Bcrypt;
+use GeebyDeeby\Crypt\PasswordHasher;
 
 use function is_object;
 
@@ -161,8 +161,8 @@ class UserController extends AbstractBase
                     'Name' => $view->fullname, 'Address' => $view->address,
                 ];
                 if (!empty($password1)) {
-                    $bcrypt = new Bcrypt();
-                    $update['Password_Hash'] = $bcrypt->create($password1);
+                    $hasher = new PasswordHasher();
+                    $update['Password_Hash'] = $hasher->create($password1);
                 }
                 $table->update(
                     $update,